Hysteresis always calculated, not only when relays are off

+++ src/stc1000p.c
@@ -357,7 +357,7 @@ void ctrl_task(void)
 
    // Start with updating the alarm
    // cache whether the 2nd probe is enabled or not.
-   probe2 = eeprom_read_config(EEADR_MENU_ITEM(Pb2)); 
+   probe2 = (uint8_t)eeprom_read_config(EEADR_MENU_ITEM(Pb2)); 
    if (ad_err1 || (ad_err2 && probe2))
    {
        sound_alarm = true;
